requirements.txt
# === CORE DATA SCIENCE ===
numpy>=1.26.4,<2.0
pandas>=2.2.3,<3.0
scipy>=1.14.0,<2.0
scikit-learn>=1.5.0,<2.0

# === FINANCIAL DATA PROVIDERS ===
databento>=0.37.0,<1.0
databento-dbn>=0.37.0,<1.0
yfinance>=0.2.37,<1.0
fredapi>=0.5.1,<1.0

# === DATABASE & STORAGE ===
duckdb>=0.10.0,<2.0
pyarrow>=14.0.0,<21.0
polars>=0.20.0,<2.0
bottleneck>=1.3.7,<2.0
numexpr>=2.8.0,<3.0

# === ASYNC & NETWORKING ===
aiohttp>=3.10.11,<4.0
requests>=2.32.0,<3.0
nest-asyncio>=1.6.0,<2.0
tenacity>=8.2.0,<10.0

# === ML/GPU (Apple Silicon Optimized) ===
mlx>=0.0.1,<1.0; sys_platform == "darwin" and platform_machine == "arm64"
# torch>=2.4.0,<3.0; sys_platform == "darwin" and platform_machine == "arm64"  # DISABLED - PyTorch fallback only
faiss-cpu>=1.7.0,<2.0  # CPU-only version optimized for Apple Silicon Metal
sentence-transformers>=2.2.0,<3.0  # Required for Einstein semantic search

# === VECTOR SEARCH & OPTIMIZATION (Apple Silicon) ===
# hnswlib>=0.8.0,<1.0  # REMOVED - FAISS is faster and more feature-rich on M4 Pro
# usearch>=2.9.0,<3.0  # REMOVED - redundant with FAISS+HNSW combination

# === OBSERVABILITY ===
opentelemetry-api>=1.22.0,<2.0
opentelemetry-sdk>=1.22.0,<2.0
opentelemetry-exporter-otlp>=1.22.0,<2.0
logfire>=3.18.0,<4.0

# === TRADING CALENDARS ===
pandas-market-calendars>=4.0.0,<6.0
exchange-calendars>=4.2.0,<5.0
quantlib>=1.33,<2.0; sys_platform == "darwin"

# === CONFIGURATION & VALIDATION ===
pydantic>=2.7.0,<3.0
pydantic-settings>=2.6.1,<3.0
python-dotenv>=1.0.1,<2.0
pyyaml>=6.0.2,<7.0

# === CLI & UI ===
click>=8.1.7,<9.0
rich>=13.7.1,<15.0
typer>=0.9.0,<1.0

# === SYSTEM & UTILITIES ===
cryptography>=45.0.4,<46.0
python-dateutil>=2.9.0,<3.0
pytz>=2024.1,<2026.0
tzdata>=2025.2,<2026.0
joblib>=1.5.1,<2.0
attrs>=25.3.0,<26.0
cffi>=1.17.1,<2.0; platform_python_implementation != "PyPy"
typing-extensions>=4.14.0,<5.0